requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
Pillow>=9.0.0
torch>=1.10.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
pycairo>=1.24.0
tqdm>=4.65.0
